requirements.txt
Flask==3.1.1
email_validator==2.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.1.1
pydenticon==0.3.1
psycopg2==2.9.10